rename: errorMessage 위치 변경
